refactor: modularize ai manager and model manager
