+++ src/ai_ml/inference/__init__.py
@@ -1,0 +1,3 @@
+from .manager import AIManager
+
+__all__ = ["AIManager"]
