+++ src/ai_ml/model_loading/__init__.py
@@ -1,0 +1,3 @@
+from .manager import ModelManager
+
+__all__ = ["ModelManager"]
